node graph settings and execution
Node types can now be defined and then added to a graph. These settings
can then be accessed by the node when the whole graph is compiled into
an executable form. To execute the graph is first saved to string form
in a dictionary of save slots and then the "brain" compiler takes the
strings as input.