Moving!
The agents now have outputs to rotation and movement (x,y,z) which
update and animate their object in Blender. A simulation class contains
the agents and has a single handler that can be registered to advance
the simulation when the frame moves.
The node editor can be launched from the blender interface and return
graphs that have been saved there (although not load them at the
moment). A start simulation button has been added which needs to be
pressed before changes to the simulation can be seen. Need to explore
automatically setting up the simulation before advancing a frame if any
settings have been changed.
A proper interpreter has been added instead of using eval() and exec().
There is one interpreter object that is reset and used again by all
nodes.