Sound, proper loading and saving and PEP8
Channels have progressed so that agents can hear each other. The GUI can
now load as well as save and all data is now properly stored in the
.blend. The data passed by nodes has been updated and the whole project
has been made PEP8 compliant. Agents now have local variables and access
variables (which include tags) that other agents (and maybe plug-ins if
a system for that is created) can view and change.