Comments! and ground channel fix
The ground channel was only calculating for the first agent that used it
each frame. Now that it is fixed the frame rate is very bad when using a
complex ground mesh (needs a way to sort the faces before ray casting).

Lots of comments have been added to cover the whole project.