Improved ground finding and hand in!
Last commit before this is handed in as my A level computing coursework!
The ground finding channel now uses a pair of binary trees (x and y
directions) to narrow down the number of faces that have to be checked
with ray casting. Currently the binary trees are very slow to generate
but lightning fast once they are done. They also assume that the ground
mesh never changes... reasonable assumption, I think. This is PN signing
off for the summer exams :P