InAIte is born!
It now has a proper name! InAIte from the word innate but with more
artificial intelligence (AI) in it.

Complete overhaul of the UI. Completely within Blenders interface now
and has no dependency on PySide. Also added prediction to the sound
channel and visual feedback on the node editor. State machines (in their
current form) are being removed and will be replaced at a later date.
The ground channel, proper collision detection and vision are being put
on hold until the official build of Blender has BVH trees exposed to
Python.