Actions preliminary work
Some ground works before actions are added. States are now highlighted
when they are the current state and the current state will return to the
start state if there isn't a valid next state. Checking the cycleState
will mean the state can return to itself without having to connect the
output to the input.